move tokenizers under data/
