+++ tests/torchtune/datasets/test_alpaca_dataset.py
@@ -11,8 +11,8 @@ from datasets import Dataset
 
 from tests.test_utils import get_assets_path
 from torchtune.data._common import CROSS_ENTROPY_IGNORE_IDX
+from torchtune.data.tokenizers import SentencePieceEncoding
 from torchtune.datasets import alpaca_cleaned_dataset, alpaca_dataset
-from torchtune.modules.tokenizers import SentencePieceEncoding
 
 
 class TestAlpacaDataset:
