+++ tests/torchtune/datasets/test_slimorca_dataset.py
@@ -9,9 +9,9 @@ import pytest
 from datasets import Dataset
 
 from tests.test_utils import get_assets_path
+from torchtune.data.tokenizers import SentencePieceEncoding
 
 from torchtune.datasets import slimorca_dataset
-from torchtune.modules.tokenizers import SentencePieceEncoding
 
 
 class TestSlimOrcaDataset:
