+++ torchtune/datasets/_samsum.py
@@ -4,8 +4,8 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from torchtune.data.tokenizers import Tokenizer
 from torchtune.datasets import instruct_dataset, InstructDataset
-from torchtune.modules.tokenizers import Tokenizer
 
 
 def samsum_dataset(
